feat: add links component
