+++ src/components/auth/auth-links/auth-links.tsx
@@ -33,56 +33,4 @@ const AuthLinks: React.FC = () => {
 	);
 };
 
-const SignInLink: React.FC = () => {
-	const { t } = useTranslation();
-
-	return (
-		<Link to={AppRoute.SIGN_IN}>
-			<Typography color="secondary.main" component="span" variant="dmSansBold">
-				{t("SignUpComponent.doYouHaveAnAccount")}
-			</Typography>
-			<Typography
-				color="primary.main"
-				component="span"
-				ml="10px"
-				variant="playfairDisplay"
-			>
-				{t("SignUpComponent.signIn")}
-			</Typography>
-		</Link>
-	);
-};
-
-const SignUpLink: React.FC = () => {
-	const { t } = useTranslation();
-
-	return (
-		<Link to={AppRoute.SIGN_UP}>
-			<Typography color="secondary.main" component="span" variant="dmSansBold">
-				{t("SignInComponent.donNotHaveAnAccountYet")}
-			</Typography>
-			<Typography
-				color="primary.main"
-				component="span"
-				ml="10px"
-				variant="playfairDisplay"
-			>
-				{t("SignUpComponent.signUp")}
-			</Typography>
-		</Link>
-	);
-};
-
-const ForgotPasswordLink: React.FC = () => {
-	const { t } = useTranslation();
-
-	return (
-		<Link to={AppRoute.FORGOT_PASSWORD}>
-			<Typography color="primary.main" variant="playfairDisplay">
-				{t("SignInComponent.forgotPassword")}
-			</Typography>
-		</Link>
-	);
-};
-
-export { AuthLinks, ForgotPasswordLink, SignInLink, SignUpLink };
+export { AuthLinks };
