+++ src/components/auth/index.ts
@@ -1,6 +1,8 @@
 export { AuthLinks } from "./auth-links/auth-links.tsx";
+export { ForgotPasswordLink } from "./forgot-password-link/forgot-password-link.tsx";
 export { GoogleButton } from "./google-button/google-button.tsx";
 export { SignInForm } from "./sign-in-form/sign-in-form.tsx";
 export { SignInLink } from "./sign-in-link/sign-in-link.tsx";
 export { SignUpForm } from "./sign-up-form/sign-up-form.tsx";
+export { SignUpLink } from "./sign-up-link/sign-up-link.tsx";
 export { StyledFormContainer } from "./styled-form-container/styles-form-container.tsx";
